Text
I'm not good in texts, but I hope I showed idea well: not to beg for help, but attract people to help. Cards need to be honest about what kind of skills will be need for certain kind of contribution - that should help route people to the right direction among the cards.